# api/openapi.yaml
openapi: 3.1.0
info:
  title: AgentOS API
  version: 0.1.0
  summary: Multi-tenant AI-agent infrastructure platform API
  description: |
    AgentOS is a multi-tenant AI-agent infrastructure platform. This contract
    covers the versioned REST API (/api/v1), the worker pull endpoint, the
    Server-Sent Events (SSE) execution-event stream, and platform health /
    metrics endpoints.

    ## Authentication
    Two schemes are accepted on every protected endpoint (see
    `securitySchemes`):
    1. `bearerToken` — `Authorization: Bearer <token>` where `<token>` is the
       HMAC-SHA256 JWT-style token returned by `POST /auth/login`
       (registration creates the first OWNER user; log in to get a token).
    2. `apiKey` — `X-API-Key: ak_...` header. API keys are created server-side
       (dev key created at startup) and authenticate as the key's OWNER role.
    Browser `EventSource` clients that cannot set headers may pass
    `?api_key=<key>` as a query parameter (legacy compatibility).

    ## RBAC
    Permissions: `agents.read`, `agents.write`, `runs.read`, `runs.execute`,
    `users.manage`, `organization.manage`. Roles: `OWNER` (all), `ADMIN`
    (all but organization.manage), `MEMBER` (agents.read, runs.read,
    runs.execute), `VIEWER` (agents.read, runs.read). Each operation declares
    its required permission via `x-required-permission`.

    ## Multi-tenancy
    Requests are scoped to an organization derived from the auth token. For
...

    ## Error model
...

...
servers:
...
security:
- bearerToken: []
- apiKey: []
paths:
...
  /eval-datasets:
    get:
      operationId: listEvalDatasets
      summary: List evaluation datasets (tenant-scoped)
      security:
      - bearerToken: []
      - apiKey: []
      tags:
      - Evaluations
      responses:
        '200':
          description: Datasets of the caller's organization, newest first.
          content:
            application/json:
              schema:
                type: object
                required:
                - datasets
                properties:
                  datasets:
                    type: array
                    items:
                      $ref: '#/components/schemas/EvalDatasetSummary'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /eval-datasets/create:
    post:
      operationId: createEvalDataset
      summary: Create an evaluation dataset with its cases
      description: |
        Validates every case before persisting: supported scorers are
        `exact`, `contains`, `regex`, `latency_under_ms`, `cost_under_cents`;
        `regex` requires `params.pattern` (compilable), `latency_under_ms`
        requires `params.threshold_ms > 0`, `cost_under_cents` requires
        `params.threshold_cents > 0`; case ids must be unique (auto-generated
        when omitted); at most 50 cases per dataset.
      security:
      - bearerToken: []
      - apiKey: []
      tags:
      - Evaluations
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
              - name
              properties:
                name:
                  type: string
                description:
                  type: string
                cases:
                  type: array
                  maxItems: 50
                  items:
                    $ref: '#/components/schemas/EvalCase'
      responses:
        '201':
          description: Created dataset (echoes the ordered case list).
          content:
            application/json:
              schema:
                type: object
                required:
                - dataset
                properties:
                  dataset:
                    $ref: '#/components/schemas/EvalDataset'
        '400':
          description: Validation error (unknown scorer, bad params, duplicate case ids, >50 cases).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EvalErrorEnvelope'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
...
  /eval-runs/compare:
    post:
      operationId: compareEvalRuns
      summary: Compare a candidate run against a baseline run
      description: |
        Cases are matched by case_id; only the intersection of both runs is
        judged. A regression is a case the baseline passed and the candidate
        failed; an improvement is the reverse.
      security:
      - bearerToken: []
      - apiKey: []
      tags:
      - Evaluations
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
              - baseline_run_id
              - candidate_run_id
              properties:
                baseline_run_id:
                  type: string
                candidate_run_id:
                  type: string
      responses:
        '200':
          description: Comparison with both summaries and case-level deltas.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EvalComparison'
        '400':
          description: Missing run ids.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EvalErrorEnvelope'
        '404':
          description: Baseline or candidate run not found in the caller's organization.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EvalErrorEnvelope'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
...
components:
...
  securitySchemes:
    bearerToken:
      type: http
      scheme: bearer
      bearerFormat: HS256-HMAC
      description: |
        HMAC-SHA256 JWT-style token from `POST /auth/login` (24h expiry).
        Claims: user_id, organization_id, email, role, exp.
    apiKey:
      type: apiKey
      in: header
      name: X-API-Key
      description: |
        Pre-provisioned API key (`ak_...`). Authenticates as the key's
        organization with the OWNER role. Browser EventSource clients may
        alternatively pass `?api_key=<key>` (legacy query-param auth).
  responses:
    Unauthorized:
      description: Missing/invalid credentials (plain text body today; JSON model is the target).
      content:
        text/plain:
          schema:
            $ref: '#/components/schemas/LegacyTextError'
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Forbidden:
      description: Authenticated but lacking the required permission or organization access.
      content:
        text/plain:
          schema:
            $ref: '#/components/schemas/LegacyTextError'
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
  schemas:
    LegacyTextError:
      type: string
      description: Legacy plain-text error body emitted via http.Error during migration.
      example: agent not found
    Error:
      type: object
      description: Structured error model (internal/httpx.ErrorModel).
      properties:
        error:
          type: object
          properties:
            code:
              type: string
              description: Machine-readable stable error code (e.g. unauthorized, forbidden, not_found,
                bad_request, internal_error).
              examples:
              - not_found
            message:
              type: string
              description: Human-readable message.
              examples:
              - agent not found
            request_id:
              type: string
              description: Correlation ID echoed from the X-Request-ID header (or generated per request).
              examples:
              - 1b8f0c47e51d2a3c4d5e6f708192a3b4
          required:
          - code
          - message
          - request_id
      required:
      - error
...
    EvalErrorEnvelope:
      type: object
      required:
      - error
      properties:
        error:
          type: object
          required:
          - code
          - message
          properties:
            code:
              type: string
              example: VALIDATION_ERROR
            message:
              type: string
    EvalParams:
      type: object
      properties:
        pattern:
          type: string
          description: Regular expression for the regex scorer.
        threshold_ms:
          type: number
          format: double
          description: Inclusive latency ceiling (ms) for latency_under_ms.
        threshold_cents:
          type: number
          format: double
          description: Inclusive cost ceiling (cents) for cost_under_cents.
    EvalCase:
      type: object
      required:
      - scorer
      properties:
        id:
          type: string
          description: Caller-supplied case id (auto-generated UUID when omitted); unique per dataset.
          example: c1
        input:
          type: string
        expected:
          type: string
          description: Expected output (exact/contains) — regex and threshold scorers use params.
        scorer:
          type: string
          enum:
          - exact
          - contains
          - regex
          - latency_under_ms
          - cost_under_cents
        params:
          $ref: '#/components/schemas/EvalParams'
    EvalDatasetSummary:
      type: object
      required:
      - id
      - name
      - case_count
      - created_at
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: string
        case_count:
          type: integer
        created_at:
          type: string
          format: date-time
    EvalDataset:
      allOf:
      - $ref: '#/components/schemas/EvalDatasetSummary'
      - type: object
        properties:
          cases:
            type: array
            items:
              $ref: '#/components/schemas/EvalCase'
...
    EvalSummary:
      type: object
      required:
      - pass_rate
      - avg_latency_ms
      - total_cost_cents
      - by_scorer
      properties:
        pass_rate:
          type: number
          format: double
          example: 0.83
        avg_latency_ms:
          type: number
          format: double
          example: 420
        total_cost_cents:
          type: number
          format: double
          example: 37
        by_scorer:
          type: object
          additionalProperties:
            type: object
            required:
            - passed
            - failed
            properties:
              passed:
                type: integer
              failed:
                type: integer
...
    EvalCaseComparison:
      type: object
      required:
      - case_id
      - baseline_passed
      - candidate_passed
      properties:
        case_id:
          type: string
        baseline_passed:
          type: boolean
        candidate_passed:
          type: boolean
    EvalComparison:
      type: object
      required:
      - baseline
      - candidate
      - regressions
      - improvements
      properties:
        baseline:
          $ref: '#/components/schemas/EvalSummary'
        candidate:
          $ref: '#/components/schemas/EvalSummary'
        regressions:
          type: array
          items:
            $ref: '#/components/schemas/EvalCaseComparison'
        improvements:
          type: array
          items:
            $ref: '#/components/schemas/EvalCaseComparison'
...